requirements.txt
torch
torchvision
numpy
Pillow
opencv-python
tqdm
scikit-learn
tensorboard
streamlit
torchsummary